backend/services/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic==2.7.1
email-validator==2.1.1
redis==5.0.4
